Review 삭제시 제거 가능한지 확인하는 기능을 확인

+++ src/main/java/com/backend/backend/service/ReviewService.java
@@ -1,7 +1,10 @@
 package com.backend.backend.service;
 
 
+import com.backend.backend.domain.Member;
 import com.backend.backend.domain.Review;
+import com.backend.backend.exception.NotExistException;
+import com.backend.backend.exception.PermissionDeniedException;
 import com.backend.backend.repository.postRepository.PostRepository;
 import com.backend.backend.repository.reviewRepository.ReviewRepository;
 import lombok.RequiredArgsConstructor;
@@ -39,10 +42,20 @@ public class ReviewService {
 
     /**
      * review를 구분 가능한 방법이 id뿐이기에 id로 검색하여 종료
+     * 해당 Member의 리뷰가 아닐 경우 삭제 불가능
      * @param id
      */
     @Transactional
-    public void deleteReview(Long id){
-        reviewRepository.deleteReviewById(id);
+    public void deleteReview(Long id, Member member){
+        Review review = reviewRepository.findReviewById(id);
+        if(review==null){
+            throw new NotExistException("존재하지 않는 리뷰 입니다");
+        }
+        if(review.getWriter().getId()== member.getId()) {
+            reviewRepository.deleteReviewById(id);
+        }
+        else {
+            throw new PermissionDeniedException("해당 멤버에게 이 리뷰를 삭제할 권한이 없습니다");
+        }
     }
 }
